+++ src/lib/components/abi/AbiForm.tsx
@@ -1,4 +1,4 @@
-import { Flex } from "@chakra-ui/react";
+import { Flex, Heading } from "@chakra-ui/react";
 import { useForm } from "react-hook-form";
 
 import type { AbiFormData, ExposedFunction } from "lib/types";
@@ -28,25 +28,35 @@ export const AbiForm = ({
   return (
     <Flex direction="column" gap={4}>
       {Object.keys(typeArgs).length > 0 && (
-        <TypesForm
-          genericTypeParams={fn.generic_type_params}
-          initialData={typeArgs}
-          propsOnChange={(value) => {
-            setValue("typeArgs", value);
-            propsOnChange?.(getValues());
-          }}
-        />
+        <Flex direction="column" gap={4}>
+          <Heading variant="h6" as="h6" color="text.main">
+            type_args
+          </Heading>
+          <TypesForm
+            genericTypeParams={fn.generic_type_params}
+            initialData={typeArgs}
+            propsOnChange={(value) => {
+              setValue("typeArgs", value);
+              propsOnChange?.(getValues());
+            }}
+          />
+        </Flex>
       )}
       {Object.keys(args).length > 0 && (
-        <ArgsForm
-          params={fn.params}
-          initialData={args}
-          propsOnChange={(value) => {
-            setValue("args", value);
-            propsOnChange?.(getValues());
-          }}
-          propsOnErrors={propsOnErrors}
-        />
+        <Flex direction="column" gap={4}>
+          <Heading variant="h6" as="h6" color="text.main">
+            args
+          </Heading>
+          <ArgsForm
+            params={fn.params}
+            initialData={args}
+            propsOnChange={(value) => {
+              setValue("args", value);
+              propsOnChange?.(getValues());
+            }}
+            propsOnErrors={propsOnErrors}
+          />
+        </Flex>
       )}
     </Flex>
   );